share the setup wizard e2e helpers and mock users
